renamed paramiko to actual used name
