+++ lib/ansible/plugins/connection/paramiko.py
